webfs: refactor tx.go
